my entire work on summariser + hybrid retrieval + advanced query system
Worked on another space due to limitations. This code file contains all the information, the defined functions and the hybrid embeddings based on distilbert model and bm25 in addition to the bart summariser. This work needs to be disassembled into other files. Don't remove/add any other logic as it works great, and if you do, please let me know.